+++ tests/ext/django/test_middleware.py
@@ -10,7 +10,8 @@ class XRayTestCase(TestCase):
 
     def setUp(self):
         django.setup()
-        xray_recorder.configure(service='test', sampling=False, context=Context())
+        xray_recorder.configure(context=Context(),
+                                context_missing='LOG_ERROR')
         xray_recorder.clear_trace_entities()
 
     def tearDown(self):